+++ cryptol-saw-core/src/Verifier/SAW/CryptolEnv.hs
@@ -128,8 +128,15 @@ data ImportVisibility
   = OnlyPublic
   | PublicAndPrivate
 
+-- | The environment for capturing the Cryptol interpreter state as well as the
+--   SAWCore translations and associated state.
+--
+--  FIXME[D]: The differences in function between this and the similar
+--   C.Env?
+
 data CryptolEnv = CryptolEnv
-  { eImports    :: [(ImportVisibility, P.Import)]           -- ^ Declarations of imported Cryptol modules
+  { eImports    :: [(ImportVisibility, P.Import)]
+                                        -- ^ Declarations of imported Cryptol modules
   , eModuleEnv  :: ME.ModuleEnv         -- ^ Imported modules, and state for the ModuleM monad
   , eExtraNames :: MR.NamingEnv         -- ^ Context for the Cryptol renamer
   , eExtraTypes :: Map T.Name T.Schema  -- ^ Cryptol types for extra names in scope
@@ -178,6 +185,12 @@ nameMatcher xs =
                        init cs == C.modNameChunksText top ++ map identText ns
 
 -- Initialize ------------------------------------------------------------------
+
+-- FIXME: Code duplication, these three functions are relatively similar (and last 2 are 85% similar):
+--  - initCryptolEnv
+--  - loadCryptolModule
+--  - importModule
+--- TODO: common up the common code.
 
 initCryptolEnv ::
   (?fileReader :: FilePath -> IO ByteString) =>
@@ -363,8 +376,10 @@ genTermEnv sc modEnv cryEnv0 = do
   let declGroups = concatMap T.mDecls
                  $ filter (not . T.isParametrizedModule)
                  $ ME.loadedModules modEnv
-  cryEnv <- C.importTopLevelDeclGroups sc C.defaultPrimitiveOptions cryEnv0 declGroups
-  traverse (\(t, j) -> incVars sc 0 j t) (C.envE cryEnv)
+      nominals   = ME.loadedNominalTypes modEnv
+  cryEnv1 <- C.genCodeForNominalTypes sc nominals cryEnv0
+  cryEnv2 <- C.importTopLevelDeclGroups sc C.defaultPrimitiveOptions cryEnv1 declGroups
+  traverse (\(t, j) -> incVars sc 0 j t) (C.envE cryEnv2)
 
 --------------------------------------------------------------------------------
 
@@ -424,7 +439,7 @@ loadCryptolModule sc primOpts env path = do
                                      (ME.loadedNominalTypes modEnv)
 
   newTermEnv <-
-    do cEnv <- C.genNominalConstructors sc newNominal oldCryEnv
+    do cEnv <- C.genCodeForNominalTypes sc newNominal oldCryEnv
        newCryEnv <- C.importTopLevelDeclGroups sc primOpts cEnv newDeclGroups
        traverse (\(t, j) -> incVars sc 0 j t) (C.envE newCryEnv)
 
@@ -508,7 +523,7 @@ importModule sc env src as vis imps = do
   m <- case mtop of
          T.TCTopModule m -> pure m
          T.TCTopSignature {} ->
-            fail "Expected a moodule but found an interface."
+            fail "Expected a module but found an interface."
   checkNotParameterized m
 
   -- Regenerate SharedTerm environment.
@@ -526,7 +541,7 @@ importModule sc env src as vis imps = do
                                      (ME.loadedNominalTypes modEnv)
 
   newTermEnv <-
-    do cEnv      <- C.genNominalConstructors sc newNominal oldCryEnv
+    do cEnv      <- C.genCodeForNominalTypes sc newNominal oldCryEnv
        newCryEnv <- C.importTopLevelDeclGroups sc C.defaultPrimitiveOptions
                                                             cEnv newDeclGroups
        traverse (\(t, j) -> incVars sc 0 j t) (C.envE newCryEnv)