+++ CHANGES.md
@@ -11,6 +11,23 @@ Note that build.sh is in any case the recommended way to build.
 
 ## New Features
 
+* SAW now supports importing Cryptol code that uses enums.
+
+  - One missing feature is supporting case default alternatives where we
+    bind the whole scrutinee, such as this example:
+
+        case s of
+           C1 ... -> EXP1
+           y      -> EXP2 -- 'y' is free in EXP2
+
+    This, more conventional default alternative, *is* supported:
+
+        case s of
+           C1 ... -> EXP1
+           _      -> EXP2 -- '_' is not used.
+
+    And case expressions that don't need a default case are supported.
+
 * SAW now ships with a `hie.yaml` file for haskell-language-server.
 
 * Add new SAWScript commands `term_eval` and `term_eval_unint`.